fix(contributions): accept fractional actual hours on evidence submit
Submitting evidence with a non-integer "Actual Hours Worked" (e.g.
6.0158) returned "invalid request body": SubmitEvidenceRequest /
Contribution.ActualDuration were Go int, so json.Decode rejected any
fractional number before the handler ran. Widen both to float64 (matching
ActualCost, which was already float64); nothing computes on duration as an
int. Frontend caps Hours and Cost to 2dp via step="0.01" + round-on-blur,
and rounds again on submit so a prefilled 6.0158 is sent as 6.02.

Also folds in the previously-requested contributions UI tweak that shares
these files: remove the "Archive Contribution" button from the detail
dialog and make the edit dialog's "Delete Contribution" button full-width
(archiving now happens only via the edit dialog).

+++ backend/internal/contributions/models.go
@@ -393,7 +393,7 @@ type Contribution struct {
 	ContributionType       ProposalType       `json:"contribution_type"`
 	Priority               Priority           `json:"priority"`
 	EstimatedDuration      int                `json:"estimated_duration"`
-	ActualDuration         int                `json:"actual_duration,omitempty"`
+	ActualDuration         float64            `json:"actual_duration,omitempty"`
 	Budget                 string             `json:"budget,omitempty"`
 	ActualCost             float64            `json:"actual_cost,omitempty"`
 	Deadline               *time.Time         `json:"deadline,omitempty"`
